+++ src/Enums/EventEnum.php
@@ -21,5 +21,7 @@ namespace vwo\Enums;
 class EventEnum {
     const VWO_VARIATION_SHOWN = 'vwo_variationShown';
     const VWO_SYNC_VISITOR_PROP = 'vwo_syncVisitorProp';
+    const VWO_ERROR = 'vwo_log';
+    const VWO_SDK_INIT_EVENT = 'vwo_fmeSdkInit';
 }
 ?>