+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.9.0] - 2025-08-05
+
+### Added
+
+- Added support for sending a one-time initialization event to the server to verify correct SDK setup.
+- Added support for sending error logs to VWO server for better debugging.
+
 ## [1.8.1] - 2025-07-24
 
 ### Added